debugging

+++ app/models/character.js
@@ -5,7 +5,7 @@ sequelize.query
 
 // Creates a "Character" model that matches up with DB
 const Character = sequelize.define("character", {
-    user: Sequelize.INTEGER,
+    user: Sequelize.BIGINT,
     data: Sequelize.JSON
 });
